changed latency field definition

+++ database/migrations/2019_01_29_184802_add_latency_to_nodes_table.php
@@ -14,7 +14,7 @@ class AddLatencyToNodesTable extends Migration
     public function up()
     {
         Schema::table('nodes', function (Blueprint $table) {
-            $table->float('latency', 8, 2)->nullable();
+            $table->decimal('latency', 5, 4)->nullable();
         });
     }
